[SDK][Typescript] Split domain modules (#3793)
