+++ .changeset/shaggy-months-post.md
@@ -1,0 +1,5 @@
+---
+"@human-protocol/sdk": patch
+---
+
+Split combined domain files into module folders with explicit files per responsibility.
