+++ packages/sdk/typescript/human-protocol-sdk/src/encryption/index.ts
@@ -1,0 +1,3 @@
+export { Encryption } from './encryption';
+export { EncryptionUtils } from './encryption_utils';
+export type { MessageDataType } from './types';
