+++ packages/sdk/typescript/human-protocol-sdk/src/escrow/index.ts
@@ -1,0 +1,2 @@
+export { EscrowClient } from './escrow_client';
+export { EscrowUtils } from './escrow_utils';
