+++ packages/sdk/typescript/human-protocol-sdk/src/kvstore/index.ts
@@ -1,0 +1,2 @@
+export { KVStoreClient } from './kvstore_client';
+export { KVStoreUtils } from './kvstore_utils';
