+++ packages/sdk/typescript/human-protocol-sdk/src/operator/index.ts
@@ -1,0 +1,1 @@
+export { OperatorUtils } from './operator_utils';
