+++ packages/sdk/typescript/human-protocol-sdk/src/staking/index.ts
@@ -1,0 +1,2 @@
+export { StakingClient } from './staking_client';
+export { StakingUtils } from './staking_utils';
