+++ packages/sdk/typescript/human-protocol-sdk/src/statistics/index.ts
@@ -1,0 +1,1 @@
+export { StatisticsUtils } from './statistics_utils';
