+++ packages/sdk/typescript/human-protocol-sdk/src/transaction/index.ts
@@ -1,0 +1,1 @@
+export { TransactionUtils } from './transaction_utils';
