+++ packages/sdk/typescript/human-protocol-sdk/src/worker/index.ts
@@ -1,0 +1,1 @@
+export { WorkerUtils } from './worker_utils';
